Optimize dynamic_path_swap colon list actions
The previous implementation was wasteful because the only available way
to get the size of an element in the colonlist was to use
clgetsize() and pass in the index of the element. clgetsize() then had
to iterate to find the element when all that was really necessary was
basically a strchr. The reason that function takes an index is because
it is useful when one does not actually have a pointer to the element
they wish to find the size of (in this case, the size of the old prefix
in the old colonlist).

This commit adds clgetsize_ptr allowing more flexibility in how one can
find the size of the element they're after.