Refactor prefix list buffers to be dynamic
Previously there was a static 1024 byte buffer that the prefix lists
would go in, but that's a stupid limitation that shouldn't exist. So
dynamically allocate those buffers. Mostly done, still need to implement
handling if dmtcp_get_reset_env() returns -2, indicating that the buffer
isn't big enough and we need to realloc